chore: reorganize directory structure and move files
