+++ api-backend/src/main/java/kr/cs/interdata/api_backend/infra/ThresholdStore.java
@@ -1,8 +1,7 @@
-package kr.cs.interdata.api_backend.dto;
+package kr.cs.interdata.api_backend.infra;
 
 import org.springframework.stereotype.Component;
 
-import java.util.Collections;
 import java.util.LinkedHashMap;
 import java.util.Map;
 import java.util.Optional;
